fix test namespaces and extend the correct testcase class
see #24
