+++ tests/Unit/EnumerationValueMethodsTest.php
@@ -1,8 +1,11 @@
 <?php
 
-namespace Sourceboat\Enumeration\Tests;
+namespace Sourceboat\Enumeration\Tests\Unit;
 
 use Sourceboat\Enumeration\Rules\EnumerationValue;
+use Sourceboat\Enumeration\Tests\FruitType;
+use Sourceboat\Enumeration\Tests\TestCase;
+use Sourceboat\Enumeration\Tests\UserRole;
 
 class EnumerationValueMethodsTest extends TestCase
 {
