+++ tests/Unit/IsTest.php
@@ -1,6 +1,9 @@
 <?php
 
-namespace Sourceboat\Enumeration\Tests;
+namespace Sourceboat\Enumeration\Tests\Unit;
+
+use Sourceboat\Enumeration\Tests\TestCase;
+use Sourceboat\Enumeration\Tests\UserRole;
 
 class IsTest extends TestCase
 {
