+++ tests/Unit/Weighted/GetMembersBetweenTest.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Sourceboat\Enumeration\Tests;
+namespace Sourceboat\Enumeration\Tests\Unit\Weighted;
 
 use Illuminate\Support\Str;
-use Orchestra\Testbench\TestCase;
+use Sourceboat\Enumeration\Tests\FruitType;
+use Sourceboat\Enumeration\Tests\TestCase;
 
-class GetMembersLessThanOrEqualToTest extends TestCase
+class GetMembersBetweenTest extends TestCase
 {
     public function setUp(): void
     {
@@ -26,7 +27,7 @@ class GetMembersLessThanOrEqualToTest extends TestCase
     }
 
     /**
-     * Data provider for the test `testIsLessThanOrEqualTo`.
+     * Data provider for the test `testGetMembersBetween`.
      *
      * @return array<mixed>
      */
@@ -35,32 +36,27 @@ class GetMembersLessThanOrEqualToTest extends TestCase
         return [
             [
                 FruitType::NUT(),
-                [
-                    Str::upper(FruitType::NUT) => FruitType::NUT(),
-                ],
+                FruitType::NUT(),
+                [],
             ],
             [
+                FruitType::NUT(),
                 FruitType::BERRY(),
-                [
-                    Str::upper(FruitType::NUT) => FruitType::NUT(),
-                    Str::upper(FruitType::BERRY) => FruitType::BERRY(),
-                ],
+                [],
             ],
             [
+                FruitType::NUT(),
                 FruitType::LEGUME(),
                 [
-                    Str::upper(FruitType::NUT) => FruitType::NUT(),
                     Str::upper(FruitType::BERRY) => FruitType::BERRY(),
-                    Str::upper(FruitType::LEGUME) => FruitType::LEGUME(),
                 ],
             ],
             [
+                FruitType::NUT(),
                 FruitType::ACCESSORY_FRUIT(),
                 [
-                    Str::upper(FruitType::NUT) => FruitType::NUT(),
                     Str::upper(FruitType::BERRY) => FruitType::BERRY(),
                     Str::upper(FruitType::LEGUME) => FruitType::LEGUME(),
-                    Str::upper(FruitType::ACCESSORY_FRUIT) => FruitType::ACCESSORY_FRUIT(),
                 ],
             ],
         ];
@@ -68,13 +64,14 @@ class GetMembersLessThanOrEqualToTest extends TestCase
 
     /**
      * @dataProvider dataProvider
-     * @param \Sourceboat\Enumeration\Tests\FruitType $member
+     * @param \Sourceboat\Enumeration\Tests\FruitType $lower
+     * @param \Sourceboat\Enumeration\Tests\FruitType $higher
      * @param array<mixed> $result
      * @return void
      */
-    public function testGetMembersLessThanOrEqualTo(FruitType $member, array $result): void
+    public function testGetMembersBetween(FruitType $lower, FruitType $higher, array $result): void
     {
-        $this->assertEquals($result, FruitType::getMembersLessThanOrEqualTo($member));
-        $this->assertEquals($result, $member->getMembersLessThanOrEqualToThis());
+        $this->assertEquals($result, FruitType::getMembersBetween($lower, $higher));
+        $this->assertEquals($result, $lower->getMembersBetweenThisAnd($higher));
     }
 }
