+++ tests/Unit/Weighted/IsEqualToTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Sourceboat\Enumeration\Tests;
+namespace Sourceboat\Enumeration\Tests\Unit\Weighted;
 
-use Orchestra\Testbench\TestCase;
+use Sourceboat\Enumeration\Tests\TestCase;
 
-class IsLessThanTest extends TestCase
+class IsEqualToTest extends TestCase
 {
     public function setUp(): void
     {
@@ -25,21 +25,21 @@ class IsLessThanTest extends TestCase
     }
 
     /**
-     * Data provider for the test `testIsLessThan`.
+     * Data provider for the test `testIsGreaterThan`.
      *
      * @return array<mixed>
      */
     public function dataProvider(): array
     {
         return [
-            [ FruitType::NUT(), FruitType::NUT(), false ],
-            [ FruitType::NUT(), FruitType::BERRY(), true ],
-            [ FruitType::NUT(), FruitType::LEGUME(), true ],
-            [ FruitType::NUT(), FruitType::ACCESSORY_FRUIT(), true ],
+            [ FruitType::NUT(), FruitType::NUT(), true ],
+            [ FruitType::NUT(), FruitType::BERRY(), false ],
+            [ FruitType::NUT(), FruitType::LEGUME(), false ],
+            [ FruitType::NUT(), FruitType::ACCESSORY_FRUIT(), false ],
             [ FruitType::ACCESSORY_FRUIT(), FruitType::NUT(), false ],
             [ FruitType::ACCESSORY_FRUIT(), FruitType::BERRY(), false ],
             [ FruitType::ACCESSORY_FRUIT(), FruitType::LEGUME(), false ],
-            [ FruitType::ACCESSORY_FRUIT(), FruitType::ACCESSORY_FRUIT(), false ],
+            [ FruitType::ACCESSORY_FRUIT(), FruitType::ACCESSORY_FRUIT(), true ],
         ];
     }
 
@@ -50,8 +50,8 @@ class IsLessThanTest extends TestCase
      * @param bool $result
      * @return void
      */
-    public function testIsLessThan(FruitType $first, FruitType $second, bool $result): void
+    public function testIsGreaterThanOrEqualTo(FruitType $first, FruitType $second, bool $result): void
     {
-        $this->assertEquals($result, $first->isLessThan($second));
+        $this->assertEquals($result, $first->isEqualTo($second));
     }
 }
