+++ tests/Unit/Weighted/IsLessThanTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Sourceboat\Enumeration\Tests;
+namespace Sourceboat\Enumeration\Tests\Unit\Weighted;
 
-use Orchestra\Testbench\TestCase;
+use Sourceboat\Enumeration\Tests\TestCase;
 
-class IsEqualToTest extends TestCase
+class IsLessThanTest extends TestCase
 {
     public function setUp(): void
     {
@@ -25,21 +25,21 @@ class IsEqualToTest extends TestCase
     }
 
     /**
-     * Data provider for the test `testIsGreaterThan`.
+     * Data provider for the test `testIsLessThan`.
      *
      * @return array<mixed>
      */
     public function dataProvider(): array
     {
         return [
-            [ FruitType::NUT(), FruitType::NUT(), true ],
-            [ FruitType::NUT(), FruitType::BERRY(), false ],
-            [ FruitType::NUT(), FruitType::LEGUME(), false ],
-            [ FruitType::NUT(), FruitType::ACCESSORY_FRUIT(), false ],
+            [ FruitType::NUT(), FruitType::NUT(), false ],
+            [ FruitType::NUT(), FruitType::BERRY(), true ],
+            [ FruitType::NUT(), FruitType::LEGUME(), true ],
+            [ FruitType::NUT(), FruitType::ACCESSORY_FRUIT(), true ],
             [ FruitType::ACCESSORY_FRUIT(), FruitType::NUT(), false ],
             [ FruitType::ACCESSORY_FRUIT(), FruitType::BERRY(), false ],
             [ FruitType::ACCESSORY_FRUIT(), FruitType::LEGUME(), false ],
-            [ FruitType::ACCESSORY_FRUIT(), FruitType::ACCESSORY_FRUIT(), true ],
+            [ FruitType::ACCESSORY_FRUIT(), FruitType::ACCESSORY_FRUIT(), false ],
         ];
     }
 
@@ -50,8 +50,8 @@ class IsEqualToTest extends TestCase
      * @param bool $result
      * @return void
      */
-    public function testIsGreaterThanOrEqualTo(FruitType $first, FruitType $second, bool $result): void
+    public function testIsLessThan(FruitType $first, FruitType $second, bool $result): void
     {
-        $this->assertEquals($result, $first->isEqualTo($second));
+        $this->assertEquals($result, $first->isLessThan($second));
     }
 }
